Update main.py
added open and save file dialogues, just packed them in to see if they wuld work, and they do!
open will select the blank pdf file, we need a fiunction that can modify a pdf in specific places, and the data of course.    save file should push it to the computer, and as a bonus, i would really like it to update the maintinance log and save on the google docs, but these are 2 different API's that will require maintinance which sucks